Theme structure updated
