+++ src/theme/theme.ts
@@ -1,0 +1,233 @@
+import { createTheme } from "@mui/material/styles";
+import type { ThemeOptions } from "@mui/material/styles";
+
+const commonTheme: ThemeOptions = {
+  typography: {
+    fontFamily: '"Inter", "Roboto", "Helvetica", "Arial", sans-serif',
+    h1: {
+      fontSize: "3rem",
+      fontWeight: 700,
+      lineHeight: 1.2,
+    },
+    h2: {
+      fontSize: "2.5rem",
+      fontWeight: 600,
+      lineHeight: 1.3,
+    },
+    h3: {
+      fontSize: "2rem",
+      fontWeight: 600,
+      lineHeight: 1.4,
+    },
+    h4: {
+      fontSize: "1.5rem",
+      fontWeight: 500,
+      lineHeight: 1.4,
+    },
+    h5: {
+      fontSize: "1.25rem",
+      fontWeight: 500,
+      lineHeight: 1.5,
+    },
+    h6: {
+      fontSize: "1rem",
+      fontWeight: 500,
+      lineHeight: 1.5,
+    },
+    body1: {
+      fontSize: "1rem",
+      lineHeight: 1.6,
+    },
+    body2: {
+      fontSize: "0.875rem",
+      lineHeight: 1.6,
+    },
+  },
+  shape: {
+    borderRadius: 12,
+  },
+  spacing: 8,
+  breakpoints: {
+    values: {
+      xs: 0,
+      sm: 600,
+      md: 900,
+      lg: 1200,
+      xl: 1536,
+    },
+  },
+  components: {
+    MuiButton: {
+      styleOverrides: {
+        root: {
+          textTransform: "none",
+          borderRadius: 12,
+          padding: "12px 24px",
+          fontSize: "1rem",
+          fontWeight: 500,
+          transition: "all 0.3s ease-in-out",
+          "&:hover": {
+            transform: "translateY(-2px)",
+            boxShadow: "0 8px 25px rgba(0,0,0,0.15)",
+          },
+        },
+      },
+    },
+    MuiCard: {
+      styleOverrides: {
+        root: {
+          borderRadius: 16,
+          boxShadow: "0 4px 20px rgba(0,0,0,0.08)",
+          transition: "all 0.3s ease-in-out",
+          "&:hover": {
+            transform: "translateY(-4px)",
+            boxShadow: "0 8px 30px rgba(0,0,0,0.15)",
+          },
+        },
+      },
+    },
+    MuiChip: {
+      styleOverrides: {
+        root: {
+          borderRadius: 20,
+          fontSize: "0.875rem",
+          fontWeight: 500,
+          transition: "all 0.2s ease-in-out",
+          "&:hover": {
+            transform: "scale(1.05)",
+          },
+        },
+      },
+    },
+    MuiPaper: {
+      styleOverrides: {
+        root: {
+          borderRadius: 12,
+        },
+      },
+    },
+    MuiAppBar: {
+      styleOverrides: {
+        root: {
+          backdropFilter: "blur(20px)",
+          backgroundColor: "rgba(255, 255, 255, 0.8)",
+          boxShadow: "0 2px 20px rgba(0,0,0,0.1)",
+        },
+      },
+    },
+  },
+};
+
+export const lightTheme = createTheme({
+  ...commonTheme,
+  palette: {
+    mode: "light",
+    primary: {
+      main: "#1976d2", // Royal Blue
+      light: "#42a5f5",
+      dark: "#1565c0",
+      contrastText: "#ffffff",
+    },
+    secondary: {
+      main: "#424242", // Dark Grey
+      light: "#757575",
+      dark: "#212121",
+      contrastText: "#ffffff",
+    },
+    background: {
+      default: "#fafafa",
+      paper: "#ffffff",
+    },
+    text: {
+      primary: "#212121",
+      secondary: "#757575",
+    },
+    grey: {
+      50: "#fafafa",
+      100: "#f5f5f5",
+      200: "#eeeeee",
+      300: "#e0e0e0",
+      400: "#bdbdbd",
+      500: "#9e9e9e",
+      600: "#757575",
+      700: "#616161",
+      800: "#424242",
+      900: "#212121",
+    },
+    success: {
+      main: "#4caf50",
+    },
+    warning: {
+      main: "#ff9800",
+    },
+    error: {
+      main: "#f44336",
+    },
+    info: {
+      main: "#2196f3",
+    },
+  },
+});
+
+export const darkTheme = createTheme({
+  ...commonTheme,
+  palette: {
+    mode: "dark",
+    primary: {
+      main: "#2196f3", // Royal Blue (lighter for dark mode)
+      light: "#64b5f6",
+      dark: "#1976d2",
+      contrastText: "#ffffff",
+    },
+    secondary: {
+      main: "#757575", // Light Grey for dark mode
+      light: "#9e9e9e",
+      dark: "#424242",
+      contrastText: "#ffffff",
+    },
+    background: {
+      default: "#121212",
+      paper: "#1e1e1e",
+    },
+    text: {
+      primary: "#ffffff",
+      secondary: "#b0b0b0",
+    },
+    grey: {
+      50: "#fafafa",
+      100: "#f5f5f5",
+      200: "#eeeeee",
+      300: "#e0e0e0",
+      400: "#bdbdbd",
+      500: "#9e9e9e",
+      600: "#757575",
+      700: "#616161",
+      800: "#424242",
+      900: "#212121",
+    },
+    success: {
+      main: "#66bb6a",
+    },
+    warning: {
+      main: "#ffa726",
+    },
+    error: {
+      main: "#ef5350",
+    },
+    info: {
+      main: "#42a5f5",
+    },
+  },
+  components: {
+    ...commonTheme.components,
+    MuiAppBar: {
+      styleOverrides: {
+        root: {
+          backdropFilter: "blur(20px)",
+          backgroundColor: "rgba(30, 30, 30, 0.8)",
+          boxShadow: "0 2px 20px rgba(0,0,0,0.3)",
+        },
+      },
+    },
+  },
+});
